+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - .gitattributes file
 - \BristolSU\Support\User\UserRepository::getById method for retrieving a user by ID
+- \BristolSU\Support\User\UserRepository::setRememberToken to set a users remember token
 
 ## [2.0.1] - (05/02/2020)